+++ playa/page.py
@@ -63,7 +63,7 @@ def _extract_text_from_obj(obj: "TextObject", vertical: bool) -> Tuple[str, floa
     chars = []
     prev_end = 0.0
     for glyph in obj:
-        x, y = glyph.textstate.glyph_offset
+        x, y = glyph.glyph_offset
         off = y if vertical else x
         # FIXME: This is a heuristic!!!
         if prev_end and off - prev_end > 0.5:
@@ -345,10 +345,8 @@ class Page:
         lines = []
         strings = []
         for text in self.texts:
-            line_matrix = text.textstate.line_matrix
-            vertical = (
-                False if text.gstate.font is None else text.gstate.font.vertical
-            )
+            line_matrix = text.line_matrix
+            vertical = False if text.gstate.font is None else text.gstate.font.vertical
             lpos = -2 if vertical else -1
             if (
                 prev_line_matrix is not None